Import a386 virtual machine by Lars Brinkhoff.
a386 is a C programming library which provides a virtual machine. The
virtual machine is an abstraction of an Intel 386 running in protected
mode. Functions in the library corresponds to privileged processor
instructions and access to hardware. The intended use for the library is
to serve as a minimal hardware abstraction layer for operating system
research.

It's imported to our CVS repository to port it on OpenBSD and then make
an OpenBSD/a386 port.